fix: User Controller Update

+++ src/main/java/com/springboot/user/controller/UserController.java
@@ -121,6 +121,17 @@ public class UserController {
     public ResponseEntity deleteUser(@PathVariable ("user-id") @Positive Long userId,
                                      @AuthenticationPrincipal CustomUserDetails customUserDetails) {
 
+        Long currentUserId = customUserDetails.getUserId();
+        log.info("현재 로그인한 사용자 ID: {}", currentUserId);
+        log.info("삭제 요청된 사용자 ID: {}", userId);
+
+        // 수정 要
+        if (!checkUserRoles.isAdmin() && !currentUserId.equals(userId)) {
+            log.error("삭제 권한 없음! Forbidden 발생");
+            throw new BusinessLogicException(ExceptionCode.USER_FORBIDDEN);
+        }
+
+
         userService.matchUserId(userId, customUserDetails);
 
         userService.deleteUser(userId);
